Module reactfs::reactor

Macros

await
Ye Olde Await Macro

stream_yield

Structs

Core
The Core of the reactor containing the event loop.

EventId
The ID of an Event

FSCallStream
Stream returned by Handle::get_fs_call() which will resolve to FSRequests.

FSRequest
A filesystem request

FSResponse
A filesystem response

FlushRequest
A block device flush request

FlushResponse
A block device flush response

FutureFlush
Future returned by Handle::flush() which will resolve to () when completed.

FutureRead
Future returned by Handle::read() which will resolve to a Vec<u8> on completion.

FutureWrite
Future returned by Handle::write() which will resolve to () when completed.

Handle
A handle to the reactor::Inner structure of reactor::Core (Weak<RefCell<Inner>>)

ReadRequest
A block device read request

ReadResponse
A block device read response

SpawnedTask
Represents a spawned task.

StreamId
The ID of a Stream

TaskId
The ID of a Task

WriteRequest
A block device write request

WriteResponse
A block device write response

Enums

BDRequest
A block device request

Event
Represents an event to be sent to the reactor's Core's main loop

FutureEvent
Represents an event which destination is a Future

StreamEvent
Represents an event which destination is a Stream