Struct reactfs::reactor::Core

pub struct Core { /* fields omitted */ }

The Core of the reactor containing the event loop.

Use Core::run() to start the reactor with a Future.

It holds a reference counter reference to Inner as Rc<RefCell<Inner>>.

Methods

impl Core

pub fn new(
bd_sender: Sender<BDRequest>,
fs_sender: Sender<FSResponse>,
receiver: Receiver<Event>
) -> Core

Creates a new reactor.

Needs:

a channel to send block device requests

a channel to send filesystem responses

a channel to receive Events: block device responses or filesystem requests

pub fn handle(&mut self) -> Handle

Returns a Handle to the Inner state of the reactor.

pub fn run<F>(self, future: F) -> Result<F::Item, F::Error> where
F: Future,

Runs a future until completion.

Return value

Returns the Result of future if any.

If future is never resolved, this function never returns