Struct reactfs::reactor::Handle

pub struct Handle { /* fields omitted */ }

A handle to the reactor::Inner structure of reactor::Core (Weak<RefCell<Inner>>)

It's used to create new I/O Futures and Streams and spawn new SpawnedTasks.

Methods

impl Handle

pub fn read(&self, offset: u64, length: u64) -> FutureRead

Reads length bytes at offset and returns a ReadFuture which resolves to Vec<u8>.

pub fn write(&self, data: Vec<u8>, offset: u64) -> FutureWrite

Writes data bytes at offset and returns a WriteFuture which resolves when the write is done`.

pub fn flush(&self) -> FutureFlush

Flush block device queue and returns a FlushFuture which resolves when the flush is done`.

pub fn recv_fs_request(&self) -> FSCallStream

Return a FSCallStream which resolves to FSRequests.

pub fn send_fs_response(&self, fs_response: FSResponse)

Sends fs_response back through the filesystem channel

pub fn spawn<F>(&self, f: F) -> TaskId where
F: Future<Item = (), Error = ()> + 'static,

Spawns a new SpawnedTask in the event loop from a Future or a Stream

Return value

This function returns the TaskId to which the Future as been assigned. It's useful to then send Events to this Future.

We don't yet provide a way to retreive the StreamId of Streams.

About panics

SpawnedTasks do not catch panics, it is the responsability of the Future to do so if needed.